requirements.txt
crewai
crewai[tools]
load_dotenv
langchain
langchain-openai
langchain-huggingface
langchain-google-genai